fix: align highlight and position
When text was inserted directly on the position of the mark (begin or end of a
highlighted text region), the recorded location stored in the marginalia was
not in sync with the highlight.

This was because of the default marker insertion type
(`marker-insertion-type`). It's by default `nil`. It needs to be `t` for all
the begin and end position of each highlight.