internal: Fix the macro expansion
All the reader-macro syntax to construct lists has been replaced with normal
syntax. As far as I can see, `seq-find` seems to be the only major macro that
relies on functions within this package.

The backtick syntax to construct a list has been replaced with `list` and
`cons` functions. So no need for the "reader macro" like `(,id ,beg ,end)`.