fix: next/prev was reversed
mapcar keeps the original order, but dolist + push reverses the order of the
resultant list. dolist was used in order to eliminate `nil` elements as mapcar
keeps it (it should, as the number of elements should be identical for
map). Now the `nil` is removed with fn `remove`; to me this more intuitive
than dolist.