intrnl: Update most functions to return t/nil
Striving to be more functional. Some functions do not return a value/list. In
this case, functions should return t if an expected operation for the args
occured in addtion to enact the (intended) side effect (e.g. change a
variable).

I don't think I am strict about it -- e.g. I left the `sort-highlights`
function mostly non-fp style -- it assumes a buffer-local variable is used,
instead of importing it as an arg, and sets the variable after sorting
distructively. It was not worth updating it -- it returns t when the side
effect is enacted (sorting distructively).