add: org-remark-delete
Delete the highlight at POINT and marginal notes for it.

This function will prompt for confirmation if there is any notes present in the
marginal notes buffer.  When the marginal notes buffer is not displayed in the
current frame, it will be temporarily displayed together with the prompt for the
user to see the notes.

If there is no notes, this function will not prompt for confirmation and will
remove the highlight and deletes the entry in the marginal notes buffer.

This command is identical with passing a universal argument to
`org-remark-remove'. "