rm: Adding Org-ID automatically to file level when file is empty
This is supposed to help bridge org-remark with org-roam.  It turns out to
clutter the backlinks because every highlight would be shown as a backlink.  The
ID should be manually added when the user decides the marginal notes are worth
backlink by adding Org-ID to turn the entry into a node.