chg: var/func names to align with GNU convention: path -> file name
Regarding user option `org-remark-notes-file-path`, the GNU convention is to
call this a "file name" rather than a "path"; "path" is only used for lists of
directories as in `load-path` (pointed out by Stefan Monnier; thank you).

To align with the GNU convention, the following changes are done.  Users should
not have to change their existing customization as the old names are aliased to
the new ones.

User option:
- org-remark-notes-file-path -> org-remark-notes-file-name

Function:
- org-remark-notes-file-path-function -> org-remark-notes-file-name-function

Private Function:
- org-remark-notes-get-file-path -> org-remark-notes-get-file-name

`org-remark-notes-file-path` and `org-remark-notes-file-path-function` are used
by existing customization, thus both explicitly made obsolete and aliased to the
new file-name equivalents.