feat: Support taking annotations in eww buffers.
EWW buffers are not associated with a file, which causes
`buffer-file-name` to return nil.

This change extracts the functionality to find the source filename
into a separate function. Earlier, `org-remark` used
`buffer-file-name` to find the source filename.

This allows us to handle `eww-mode` within the new function (and
potentially more such modes in the future).

We also add `org-remark-auto-on` to the `eww-after-render-hook` to
make sure that existing annotations are loaded.

The end result is that it is now possible to annotate web-pages inside
EWW.

Function introduced:

1. `org-remark-source-find-file-name` - All direct calls to
`buffer-file-name` are now replaced by this function.

Functions changed:
1. `org-remark-global-tracking-mode` - `org-remark-auto-on` runs
on EWW render completion.
2. `org-remark-highlight-get-title` - This function can also extract
the highlight title from the EWW URL now.