feat: extend support for non-file-visiting buffer
Features:

- For non-file-visiting buffer, now the default note file name is "marginalia.org" *in the user emacs directory*: e.g. ~/.config/emacs/marginaliaorg, ~/emacs.d/marginalia.org, etc.

- For non-file-visiting buffer and if it is not viewed with EWW,
Org-remark now assumes the source name is the buffer's name.

Description:

This is continuation of the stream of fetures that EWW support opened
up.  For buffers that are not visiting a file in general, default
functions to find the notes and source files need adjustment. This is
especially important when 'org-remark-notes-file-name-function' is
selected from the user option. This assumes FILE exsits and creates a
note FILE-notes.org. For non-file-visiting buffers, now default note
file name is marginalia.org in the user emacs directory.

The treatment of source file's name requires more work. Returning the
buffer's name is not really useful (e.g. "*Info*" for all the nodes). I
believe it reqiures a handler specific to each context.

Changed file and functions:

* org-remark-global-tracking.el
(org-remark-notes-file-name-function):
(org-remark-notes-get-file-name):
(org-remark-source-find-file-name):