fix: notes-update-source to allow for beg end adjustment
Added a new function org-remark-find-overlay-in. Previously
org-remark-find-overlay-at was used. This had a problem when users
manually move the highlight in the notes by adjusting :org-remark-beg:
and :org-remark-end: properties. When the beg point does not have the
hightlight overlay, the update would fail.

Using org-remark-find-overlay-in is not perfect as the overlay will need
to be found in the location between beg and end; however, this is deemed
to be an edge case -- you would not really move the highlight.