refactor -notes-* functions
Especially the sync function has been refactored.

'org-remark-save' has been refactored to cator to the situation when the
source buffer is also the notes buffer, thus simplifies the
-highlight-save function -- longer saves the notes buffer for each
highlight. The notes buffer is now saved by 'org-remark-save' on
'after-save-buffer' hook -- this means unless the user saves the source
buffer, the notes buffer will not get saved.