fix: highlights housekeep to delete notes when you want to keep
In working toward supporting nov.el, it was found out that marginal
notes were removed when they should be kept intact.  It was found that
'-highlights-housekeep' function was the cause.  This was because nov.el
buffer becomes writable when we turn the
page:(e.g. 'nov-next-document'), erase the buffer, and then render the
new page.  Housekeeping should be done to the highlights overlays from
the now-the-previous page (as they will have start/end both 1); however,
the related notes must not be removed.

This can be prevented by adding additional condition that the major mode
is derived from special mode (nov-mode is).