fix: Exclude pdfs from default file type #57
The fix is applied to more generic case, not specific to PDF files.
Unlike epub files via nov-mode or websites via eww-mode, pdf-view-mode
buffers are actually visiting buffers. And yet, it is also derived from
special-mode, which means the buffers visited are meant to be read-only.

The fix implemented is to add a check to
'org-remark-source-find-file-name'. It now does the following:

1. Check whether or not the source buffer visited is derived from
special-mode.

2. If so, its notes buffer should be found by a mode-specific function,
not passing buffer-file-name