fix(adjust-position) newline char prevent string match
Cater to the case when the text is divided by a newline character \n.
Remove all spaces and newline chars before comparing the strings.

When the highlighted text is long and include a newline, it's not
easy (possilbe?) to determine it should be replaced with a space or no
space. Eliminate the ambiguity by removing all spaces and newline chars.