refactor(load): make load process robust.
When nov opens a epub file, if the rendering fails, the book may be
overridden by a text file of a single chapter inside the whole epub
archive, rendering the epub book useless.

This refactor makes the loading highlights by Org-remark robust so that
the problem above will be unlikely caused by Org-remark.