fix: Add headlines in the correct order and in the correct headline
This was missed in the refctoring for nov.el. It was not easy to detect
when one single notes file is created for a single source (e.g. a signle
EPUB book).

The problem was in the new logic to find and add new headlines. For each
headline, the point needs to move to inside the current headline;
otherwise, the narrowing and `point-max` would not work for the intended
headline, thus creating a new highlight headline in the wrong parent
node.

This fix adds the `goto-char` to move to the correct point in the
current headline intended to be narrowed to.