fix(change): force trigger of notes buffer face
Problem: changing to the same line-highlight does not trigger sync but
this causes the icon for notes not being updated.

Force tiggering the update save for :change:operation. The line-icons do
not get updated because :change: to the same pen does not involve buffer
modificaiton and thus the sync does not get triggered to update icons.