fix(line): redraw instead of reload when window size changes
Reload causes some troubles. An example case:

1. the buffer contains a range-highlight and a line-highlight.
2. While you are editing the buffer, you move both highlights.
3. Before saving, you press C-h e to show echo or M-x to show minibuffer

3 above triggers `window-size-change-function` and reload all the
highlights for the buffer. Since the location of the range-highlight has
changed from that saved in the marginalia.org, `position-adjusted` gets
triggered, showing the icon (when enabled). This is false. The
save-buffer would have saved the new location and the
`position-adjusted` would not have been applied.

The line-highlight also gets moved to the location saved in the
marginalia.org. Again, this should not happen -- either wait for the
user to save the buffer or undo the changes to get back where the
highlight should be.

Redrawing avoids these issues. We are now only looking at the
line-highlights, and changing the spacer overlays. This way, we can
avoid triggering `after-load` functions (including `position-adjusted`).