fix(line):when line-icon is an image, it fails to carry face
The notes icon from `org-remark-icon` package was made available to
`org-remrk-line`; however, it was assumed that the highlight the icon is
added to is an overlay added to the main text. This assumption does not
hold for line-highlights available with `org-remark-line`. It is likely
that the user customize the glyph to be displayed in the margin to be
an SVG icon. In this case, the notes icon (the icon that is added to the
highlight when the user has written an annotation to it) is added onto
another overlay to another SVG icon. This fix caters to this case, and
let the notes icon carry the same face as the underlying line-highlight
overlay.

LIMITATION made clear: if the user wants to use image icons (e.g. SVG
image icon created with package `icons', available Emacs 29.1 or
higher), it is limited to a single character with no space before and
after the character. This limitation does not apply to string of
characters without images, but it is generally assumed that the the
value set to this customizing variable will be a short string (e.g 3
characters long with a pair of parentheses before and after a single
character, such as the default value of icon customizing variables.