fix: #70 org 9.6.7 breaks org-remark-highlights-get
Reported by @alan-w-255 on GitHub. Thank you.

Relying on the fact `org-next-visible-heading` currently returns nil
when moved to a heading and returns non-nil when not (e.g. at the end of
the visible part of buffer) does not seem to be a good practice -- this
was the case before this fix.

Adding a check with `org-at-heading-p` conveys the intent of the code
and should be safe for the change of returned value of
`org-next-visible-heading` in the upcoming org version 9.6.7.