fix(line): `org-remark-mark-line` does not autoload
The comment is created via a call to `org-remark-create` macro. The call
is done within the library, so we would need to wait for it to be
loaded. This is too late, so adding a dummy `defun` with the autoload
cookie to register it. The actual implementation is added when the
library is later loaded (likely together with `org-remark` itself
as it is likely to be deferred).