fix: Highlights overlay can get added to unsupported buffer
Before this fix, a highlight could be added to an unsupported buffer
such as *scratch* buffer. When you try to remove it, you get an error
message. This is all confusing.

Now with this fix, trying to add a highlight in an unnsupported buffer,
you only get a message and the overlay is not to be added.